+++ src/gpio/convert.rs
@@ -112,14 +112,6 @@ impl<const P: char, const N: u8, MODE: PinMode> Pin<P, N, MODE> {
         self.into_mode()
     }
 
-    /// Configures the pin as a pin that can change between input
-    /// and output without changing the type. It starts out
-    /// as a floating input
-    pub fn into_dynamic(self) -> DynamicPin<P, N> {
-        self.into_floating_input();
-        DynamicPin::new(Dynamic::InputFloating)
-    }
-
     /// Puts `self` into mode `M`.
     ///
     /// This violates the type state constraints from `MODE`, so callers must
@@ -166,19 +158,19 @@ macro_rules! change_mode {
 }
 use change_mode;
 
-use super::ErasedPin;
-impl<MODE: PinMode> ErasedPin<MODE> {
+use super::AnyPin;
+impl<MODE: PinMode> AnyPin<MODE> {
     #[inline(always)]
     pub(super) fn mode<M: PinMode>(&mut self) {
         let n = self.pin_id();
-        change_mode!(self.block(), n);
+        change_mode!((*self.block()), n);
     }
 
     #[inline(always)]
     /// Converts pin into specified mode
-    pub fn into_mode<M: PinMode>(mut self) -> ErasedPin<M> {
+    pub fn into_mode<M: PinMode>(mut self) -> AnyPin<M> {
         self.mode::<M>();
-        ErasedPin::from_pin_port(self.into_pin_port())
+        AnyPin::from_pin_port(self.into_pin_port())
     }
 }