+++ src/gpio/hal_02.rs
@@ -1,8 +1,10 @@
 use core::convert::Infallible;
 
+use crate::gpio::DynamicPin;
+
 use super::{
-    dynamic::PinModeError, marker, DynamicPin, ErasedPin, Input, OpenDrain, Output,
-    PartiallyErasedPin, Pin, PinMode, PinState,
+    dynamic::PinModeError, marker, AnyPin, Input, OpenDrain, Output, PartiallyErasedPin, Pin,
+    PinMode, PinState,
 };
 
 use embedded_hal_02::digital::v2::{
@@ -107,7 +109,7 @@ where
 
 // Implementations for `ErasedPin`
 
-impl<MODE> OutputPin for ErasedPin<Output<MODE>> {
+impl<MODE> OutputPin for AnyPin<Output<MODE>> {
     type Error = core::convert::Infallible;
 
     #[inline(always)]
@@ -123,7 +125,7 @@ impl<MODE> OutputPin for ErasedPin<Output<MODE>> {
     }
 }
 
-impl<MODE> StatefulOutputPin for ErasedPin<Output<MODE>> {
+impl<MODE> StatefulOutputPin for AnyPin<Output<MODE>> {
     #[inline(always)]
     fn is_set_high(&self) -> Result<bool, Self::Error> {
         Ok(self.is_set_high())
@@ -135,7 +137,7 @@ impl<MODE> StatefulOutputPin for ErasedPin<Output<MODE>> {
     }
 }
 
-impl<MODE> ToggleableOutputPin for ErasedPin<Output<MODE>> {
+impl<MODE> ToggleableOutputPin for AnyPin<Output<MODE>> {
     type Error = Infallible;
 
     #[inline(always)]
@@ -145,7 +147,7 @@ impl<MODE> ToggleableOutputPin for ErasedPin<Output<MODE>> {
     }
 }
 
-impl<MODE> InputPin for ErasedPin<MODE>
+impl<MODE> InputPin for AnyPin<MODE>
 where
     MODE: marker::Readable,
 {