+++ src/spi.rs
@@ -957,7 +957,7 @@ impl<SPI: Instance> DmaBuilder<SPI> {
 unsafe impl<SPI: Instance> PeriAddress for Rx<SPI> {
     #[inline(always)]
     fn address(&self) -> u32 {
-        unsafe { (*SPI::ptr()).dr().as_ptr() as u32 }
+        unsafe { (*SPI::PTR).dr().as_ptr() as u32 }
     }
 
     type MemSize = u8;
@@ -971,7 +971,7 @@ unsafe impl<SPI, STREAM, const CHANNEL: u8> DMASet<STREAM, CHANNEL, PeripheralTo
 unsafe impl<SPI: Instance> PeriAddress for Tx<SPI> {
     #[inline(always)]
     fn address(&self) -> u32 {
-        unsafe { (*SPI::ptr()).dr().as_ptr() as u32 }
+        unsafe { (*SPI::PTR).dr().as_ptr() as u32 }
     }
 
     type MemSize = u8;